+++ bundles/io/org.openhab.io.dropbox/src/main/java/org/openhab/io/dropbox/internal/DropboxSyncMode.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2010-2016 by the respective copyright holders.
+ * Copyright (c) 2010-2017 by the respective copyright holders.
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
@@ -9,15 +9,13 @@
 package org.openhab.io.dropbox.internal;
 
 /**
- * Enumerates all valid synchronization Modes
+ * Enumerates all valid synchronization Modes.
  *
  * @author Thomas.Eichstaedt-Engelen
  * @since 1.0.0
  */
 public enum DropboxSyncMode {
-
     DROPBOX_TO_LOCAL,
     LOCAL_TO_DROPBOX,
     BIDIRECTIONAL;
-
 }
