+++ bundles/io/org.openhab.io.dropbox/src/main/java/org/openhab/io/dropbox/internal/DropboxSynchronizer.java
@@ -61,9 +61,9 @@ import com.dropbox.core.DbxWebAuthNoRedirect;
 import com.dropbox.core.DbxWriteMode;
 
 /**
- * The {@link DropboxSynchronizer} is able to synchronize contents of your Dropbox
+ * The {@link DropboxSynchronizer} is able to synchronize the contents of Dropbox
  * to the local file system and vice versa. There are three synchronization modes
- * available: local to dropbox (which is the default mode), dropbox to local, and
+ * available: local to dropbox (the default mode), dropbox to local, and
  * bidirectional.
  *
  * @author Thomas.Eichstaedt-Engelen - Initial Contribution
@@ -71,192 +71,93 @@ import com.dropbox.core.DbxWriteMode;
  * @author Chris Carman - add support for personalAccessToken
  * @since 1.0.0
  */
-public class DropboxSynchronizer implements ManagedService {
+public class DropboxSynchronizer {
 
-    private static final Logger logger = LoggerFactory.getLogger(DropboxSynchronizer.class);
+    private Logger logger;
 
-    private static final String DROPBOX_SCHEDULER_GROUP = "Dropbox";
-    private static final String FIELD_DELIMITER = "@@";
-    private static final String LINE_DELIMITER = System.getProperty("line.separator");
-    private static final String DELTA_CURSOR_FILE_NAME = File.separator + "deltacursor.dbx";
-    private static final String DROPBOX_ENTRIES_FILE_NAME = File.separator + "dropbox-entries.dbx";
-    private static final String AUTH_FILE_NAME = File.separator + "authfile.dbx";
+    private final String FIELD_DELIMITER = "@@";
+    private final String LINE_DELIMITER = System.getProperty("line.separator");
+    private final String DELTA_CURSOR_FILE_NAME = File.separator + "deltacursor.dbx";
+    private final String DROPBOX_ENTRIES_FILE_NAME = File.separator + "dropbox-entries.dbx";
+    private final String AUTH_FILE_NAME = File.separator + "authfile.dbx";
+
+    /** The default directory to which to download files from Dropbox */
+    private final String DEFAULT_CONTENT_DIR = DropboxUtils.getConfigDirFolder();
+
+    private final List<String> DEFAULT_UPLOAD_FILE_FILTER = Arrays.asList("^([^/]*/){1}[^/]*$",
+            "/configurations.*", "/logs/.*", "/etc/.*");
+    private final List<String> DEFAULT_DOWNLOAD_FILE_FILTER = Arrays.asList("^([^/]*/){1}[^/]*$",
+            "/configurations.*");
 
     /**
      * Holds the id of the last synchronisation cursor. This is needed to
      * define the delta to download from Dropbox.
      */
-    private static String lastCursor = null;
-    private static String lastHash = null;
-
-    //// Authentication: user must configure either the personalAccessToken or
-    //// BOTH the AppKey AND the AppSecret; if all 3 are defined, then the
-    //// personalAccessToken will be used and the others ignored.
+    private String lastCursor = null;
+    private String lastHash = null;
 
     /**
-     * A user's personal access token retrieved from configuration only;
-     * null by default
+     * The base directory to synchronize with openHAB.
      */
-    private static String personalAccessToken;
-
-    /// AppKey and AppSecret:
-    /// These are legacy attributes from when there was an official openHAB Dropbox app.
-    /// These should not generally be used, as they are more difficult to set up than
-    /// the newer method via personalAccessToken.
-    ///
-    /// Note: When using these two attributes, the user must watch the logfile during
-    /// the startup of OpenHAB to get the URL to open in a Browser to allow the plugin
-    /// to connect to a predefined App-Folder (see
-    /// <a href="https://www.dropbox.com/developers/apps">Dropbox Documentation</a>
-    /// for more information).
-
-    /** The configured AppKey (optional; see notes above) */
-    private static String appKey = "NotARealKey";
-
-    /** The configured AppSecret (optional; see notes above) */
-    private static String appSecret = "NotARealSecret";
-
-    /** The default directory to download files from Dropbox to (currently '.') */
-    private static final String DEFAULT_CONTENT_DIR = getConfigDirFolder();
-
-    /**
-     * The base directory to synchronize with openHAB (defaults to
-     * DEFAULT_CONTENT_DIR)
-     */
-    private static String contentDir = DEFAULT_CONTENT_DIR;
+    private String contentDir;
 
     /** The base directory for the .dbx files */
-    public final static String DBX_FOLDER = getUserDbxDataFolder();
-
-    /** The configured synchronization mode (defaults to LOCAL_TO_DROPBOX) */
-    private static DropboxSyncMode syncMode = DropboxSyncMode.LOCAL_TO_DROPBOX;
-
-    /**
-     * The upload interval as a Cron Expression (optional; defaults to
-     * '0 0 2 * * ?', which means once a day at 2 a.m.) */
-    private static String uploadInterval = "0 0 2 * * ?";
-
-    /**
-     * The download interval as a Cron Expression (optional; defaults to
-     * '0 0/5 * * * ?', which means every 5 minutes)
-     */
-    private static String downloadInterval = "0 0/5 * * * ?";
-
-    private static final List<String> DEFAULT_UPLOAD_FILE_FILTER = Arrays.asList("^([^/]*/){1}[^/]*$",
-            "/configurations.*", "/logs/.*", "/etc/.*");
-    private static final List<String> DEFAULT_DOWNLOAD_FILE_FILTER = Arrays.asList("^([^/]*/){1}[^/]*$",
-            "/configurations.*");
+    public final String DBX_FOLDER = DropboxUtils.getUserDbxDataFolder();
 
     /**
      * Defines a comma separated list of regular expressions which matches the
      * filenames to upload to Dropbox (optional; defaults to '/configurations.*,
      * /logs/.*, /etc/.*')
      */
-    private static List<String> uploadFilterElements = DEFAULT_UPLOAD_FILE_FILTER;
+    private List<String> uploadFilterElements = DEFAULT_UPLOAD_FILE_FILTER;
 
     /**
      * Defines a comma separated list of regular expressions which matches the
      * filenames to download from Dropbox (optional; defaults to '/configurations.*')
      */
-    private static List<String> downloadFilterElements = DEFAULT_DOWNLOAD_FILE_FILTER;
+    private List<String> downloadFilterElements = DEFAULT_DOWNLOAD_FILE_FILTER;
 
     /**
      * Operates the Synchronizer in fake mode which avoids sending files to or
      * from Dropbox. This is meant as a test mode for the filter settings.
      * (optional; defaults to false)
      */
-    private static boolean fakeMode = false;
+    private boolean fakeMode;
 
-    private static boolean isProperlyConfigured = false;
-
-    private static DropboxSynchronizer instance = null;
-
-    private static final DbxAppInfo appInfo =
-        new DbxAppInfo(DropboxSynchronizer.appKey, DropboxSynchronizer.appSecret);
-
-    private final static DbxRequestConfig requestConfig = new DbxRequestConfig("openHAB/1.0",
-            Locale.getDefault().toString());
-
-    public void activate() {
-        DropboxSynchronizer.instance = this;
+    protected DropboxSynchronizer() {
+        logger = LoggerFactory.getLogger(DropboxSynchronizer.class);
     }
 
-    public void deactivate() {
-        logger.debug("About to shut down Dropbox Synchronizer ...");
-
-        cancelAllJobs();
-        isProperlyConfigured = false;
-
-        lastCursor = null;
-        uploadFilterElements = DEFAULT_UPLOAD_FILE_FILTER;
-        downloadFilterElements = DEFAULT_DOWNLOAD_FILE_FILTER;
-
-        DropboxSynchronizer.instance = null;
-
-        logger.debug("Shutdown completed.");
+    protected void setFakeMode(boolean value) {
+        fakeMode = value;
     }
 
-    private void activateSynchronizer() {
-        if (isAuthenticated()) {
-            startSynchronizationJobs();
-        } else {
-            try {
-                startAuthentication();
-            } catch (DbxException e) {
-                logger.warn("Couldn't start authentication process: {}", e.getMessage());
-            }
+    protected void setContentDir(String value) {
+        if (StringUtils.isBlank(value) ) {
+            contentDir = DEFAULT_CONTENT_DIR;
+        }
+        else {
+            contentDir = value;
+        }
+        logger.debug("contentdir: {}", contentDir);
+    }
+
+    protected void setUploadFilterElements(List<String> value) {
+        if (value != null) {
+            uploadFilterElements = value;
+        }
+        else {
+            uploadFilterElements = DEFAULT_UPLOAD_FILE_FILTER;
         }
     }
 
-    /**
-     * Starts the OAuth authorization process with Dropbox. This is a
-     * multi-step process which is described in the Wiki.
-     * 
-     * @throws DbxException if there are technical or application level errors
-     *             in the Dropbox communication
-     * 
-     * @see <a href="https://github.com/openhab/openhab/wiki/Dropbox-IO">openHAB Dropbox IO Wiki</a>
-     */
-    public void startAuthentication() throws DbxException {
-        if (personalAccessToken == null) {
-            DbxWebAuthNoRedirect webAuth = new DbxWebAuthNoRedirect(requestConfig, appInfo);
-            String authUrl = webAuth.start();
-
-            logger.info("#########################################################################################");
-            logger.info("# Dropbox Integration: U S E R   I N T E R A C T I O N   R E Q U I R E D !!");
-            logger.info("# 1. Open URL '{}'", authUrl);
-            logger.info("# 2. Allow openHAB to access Dropbox");
-            logger.info("# 3. Paste the authorisation code here using the command 'finishAuthentication \"<token>\"'");
-            logger.info("#########################################################################################");
-        } else {
-            logger.info("#########################################################################################");
-            logger.info("# Starting auth using personal access token");
-            logger.info("#########################################################################################");
-            writeAccessToken(personalAccessToken);
-            startSynchronizationJobs();
+    protected void setDownloadFilterElements(List<String> value) {
+        if (value != null) {
+            downloadFilterElements = value;
         }
-    }
-
-    /**
-     * Finishes the OAuth authorization process by taking the given {@code token} and creating
-     * an accessToken out of it. The authorization process is a multi-step process which is
-     * described in the Wiki in detail.
-     * 
-     * @throws DbxException if there are technical or application level errors
-     *             in the Dropbox communication
-     * 
-     * @see <a href="https://github.com/openhab/openhab/wiki/Dropbox-IO">openHAB Dropbox IO Wiki</a>
-     */
-    public void finishAuthentication(String code) throws DbxException {
-        DbxWebAuthNoRedirect webAuth = new DbxWebAuthNoRedirect(requestConfig, appInfo);
-        String accessToken = webAuth.finish(code).accessToken;
-        writeAccessToken(accessToken);
-
-        logger.info("#########################################################################################");
-        logger.info("# OAuth2 authentication flow has been finished successfully ");
-        logger.info("#########################################################################################");
-
-        startSynchronizationJobs();
+        else {
+            downloadFilterElements = DEFAULT_DOWNLOAD_FILE_FILTER;
+        }
     }
 
     /**
@@ -505,7 +406,7 @@ public class DropboxSynchronizer implements ManagedService {
                 collectLocalEntries(localEntries, file.getPath());
             } else {
                 // if we are on a Windows filesystem we need to change the separator for dropbox
-                if (isWindows()) {
+                if (DropboxUtils.isWindows()) {
                     normalizedPath = normalizedPath.replace('\\', '/');
                 }
                 localEntries.put(normalizedPath, file.lastModified());
@@ -547,64 +448,21 @@ public class DropboxSynchronizer implements ManagedService {
         }
     }
 
-    private void writeAccessToken(String content) {
-        // create folder for .dbx files if it does not exist
-        File folder = new File(DBX_FOLDER);
-        if (!folder.exists()) {
-            folder.mkdirs();
-        }
-        File tokenFile = new File(DBX_FOLDER + AUTH_FILE_NAME);
-        writeLocalFile(tokenFile, content);
-    }
-
-    private String readAccessToken() {
-        File tokenFile = new File(DBX_FOLDER + AUTH_FILE_NAME);
-        return readFile(tokenFile);
-    }
-
-    private boolean isAuthenticated() {
-        return StringUtils.isNotBlank(readAccessToken());
-    }
-
     private void writeDeltaCursor(String deltaCursor) {
         if (!deltaCursor.equals(lastCursor)) {
             logger.trace("Delta-Cursor changed (lastCursor '{}', newCursor '{}')", lastCursor, deltaCursor);
             File cursorFile = new File(DBX_FOLDER + DELTA_CURSOR_FILE_NAME);
-            writeLocalFile(cursorFile, deltaCursor);
+            DropboxUtils.writeLocalFile(cursorFile, deltaCursor);
             lastCursor = deltaCursor;
         }
     }
 
     private String readDeltaCursor() {
         File cursorFile = new File(DBX_FOLDER + DELTA_CURSOR_FILE_NAME);
-        return readFile(cursorFile);
+        return DropboxUtils.readFile(cursorFile);
     }
 
-    private String readFile(File file) {
-        String content = null;
-        if (file.exists()) {
-            try {
-                List<String> lines = FileUtils.readLines(file);
-                if (lines.size() > 0) {
-                    content = lines.get(0);
-                }
-            } catch (IOException ioe) {
-                logger.debug("Handling of cursor file threw an Exception", ioe);
-            }
-        }
-        return content;
-    }
-
-    private static void writeLocalFile(File file, String content) {
-        try {
-            FileUtils.writeStringToFile(file, content);
-            logger.debug("Created file '{}' with content '{}'", file.getAbsolutePath(), content);
-        } catch (IOException e) {
-            logger.error("Couldn't write to file '{}'.", file.getPath(), e);
-        }
-    }
-
-    private static void deleteLocalFile(String fqPath) {
+    private void deleteLocalFile(String fqPath) {
         File fileToDelete = new File(fqPath);
         if (!fileToDelete.isDirectory()) {
             boolean success = true;
@@ -620,256 +478,5 @@ public class DropboxSynchronizer implements ManagedService {
         } else {
             logger.trace("Local item '{}' wasn't deleted because it is a directory.");
         }
-    }
-
-    @SuppressWarnings("rawtypes")
-    @Override
-    public void updated(Dictionary config) throws ConfigurationException {
-        if (config == null) {
-            logger.debug("Updated() was called with a null config!");
-            return;
-        }
-
-        isProperlyConfigured = false;
-
-        String appKeyString = Objects.toString(config.get("appkey"), null);
-        if (isNotBlank(appKeyString)) {
-            DropboxSynchronizer.appKey = appKeyString;
-        }
-
-        String appSecretString = Objects.toString(config.get("appsecret"), null);
-        if (isNotBlank(appSecretString)) {
-            DropboxSynchronizer.appSecret = appSecretString;
-        }
-
-        String pat = Objects.toString(config.get("personalAccessToken"), null);
-        if (isNotBlank(pat)) {
-            DropboxSynchronizer.personalAccessToken = pat;
-        }
-
-        if (logger.isDebugEnabled()) {
-            StringBuffer message = new StringBuffer();
-            message.append("Authentication parameters to be used:\r\n");
-            if (isNotBlank(pat)) {
-                message.append("     Personal access token = " + pat + "\r\n");
-            } else {
-                message.append("     appkey = " + appKeyString + "\r\n");
-                message.append("  appsecret = " + appSecretString + "\r\n");
-            }
-            logger.debug(message.toString());
-        }
-
-        if (isBlank(pat) && (isBlank(appKeyString) || isBlank(appSecretString))) {
-            throw new ConfigurationException("dropbox:authentication",
-                    "The Dropbox authentication parameters are incorrect!  The parameter 'personalAccesstoken' must be set, or both of the parameters 'appkey' and 'appsecret' must be set. Please check your configuration.");
-        }
-
-        String fakeModeString = Objects.toString(config.get("fakemode"), null);
-        if (isNotBlank(fakeModeString)) {
-            DropboxSynchronizer.fakeMode = BooleanUtils.toBoolean(fakeModeString);
-        }
-
-        String contentDirString = Objects.toString(config.get("contentdir"), null);
-        if (isNotBlank(contentDirString)) {
-            DropboxSynchronizer.contentDir = contentDirString;
-        }
-        logger.debug("contentdir: {}", contentDir);
-
-        String uploadIntervalString = Objects.toString(config.get("uploadInterval"), null);
-        if (isNotBlank(uploadIntervalString)) {
-            DropboxSynchronizer.uploadInterval = uploadIntervalString;
-        }
-
-        String downloadIntervalString = Objects.toString(config.get("downloadInterval"), null);
-        if (isNotBlank(downloadIntervalString)) {
-            DropboxSynchronizer.downloadInterval = downloadIntervalString;
-        }
-
-        String syncModeString = Objects.toString(config.get("syncmode"), null);
-        if (isNotBlank(syncModeString)) {
-            try {
-                DropboxSynchronizer.syncMode = DropboxSyncMode.valueOf(syncModeString.toUpperCase());
-            } catch (IllegalArgumentException iae) {
-                throw new ConfigurationException("dropbox:syncmode", "Unknown SyncMode '" + syncModeString
-                        + "'. Valid SyncModes are 'DROPBOX_TO_LOCAL', 'LOCAL_TO_DROPBOX' and 'BIDIRECTIONAL'.");
-            }
-        }
-
-        String uploadFilterString = Objects.toString(config.get("uploadfilter"), null);
-        if (isNotBlank(uploadFilterString)) {
-            String[] newFilterElements = uploadFilterString.split(",");
-            uploadFilterElements = Arrays.asList(newFilterElements);
-        }
-
-        String downloadFilterString = Objects.toString(config.get("downloadfilter"), null);
-        if (isNotBlank(downloadFilterString)) {
-            String[] newFilterElements = downloadFilterString.split(",");
-            downloadFilterElements = Arrays.asList(newFilterElements);
-        }
-
-        // we got this far, so we define this synchronizer as properly configured ...
-        isProperlyConfigured = true;
-        logger.debug("Bundle is properly configured. Activating synchronizer.");
-        activateSynchronizer();
-    }
-
-    // ****************************************************************************
-    // Synchronisation Jobs
-    // ****************************************************************************
-
-    private void startSynchronizationJobs() {
-        if (isProperlyConfigured) {
-            cancelAllJobs();
-            if (isAuthenticated()) {
-                logger.debug("Authenticated. Scheduling jobs.");
-                scheduleJobs();
-            } else {
-                logger.debug("Dropbox bundle isn't authorized properly, so the synchronization jobs "
-                        + "won't be started! Please re-initiate the authorization process by restarting the "
-                        + "Dropbox bundle through the OSGi console.");
-            }
-        }
-    }
-
-    /**
-     * Schedules the quartz synchronization according to the synchronization mode
-     */
-    private void scheduleJobs() {
-        switch (syncMode) {
-            case DROPBOX_TO_LOCAL:
-                logger.debug("Scheduling DROPBOX_TO_LOCAL download interval: {}", DropboxSynchronizer.downloadInterval);
-                schedule(DropboxSynchronizer.downloadInterval, false);
-                break;
-            case LOCAL_TO_DROPBOX:
-                logger.debug("Scheduling LOCAL_TO_DROPBOX upload interval: {}", DropboxSynchronizer.uploadInterval);
-                schedule(DropboxSynchronizer.uploadInterval, true);
-                break;
-            case BIDIRECTIONAL:
-                logger.debug("Scheduling BIDIRECTIONAL download interval: {}, upload interval: {}",
-                        DropboxSynchronizer.downloadInterval, DropboxSynchronizer.uploadInterval);
-                schedule(DropboxSynchronizer.downloadInterval, false);
-                schedule(DropboxSynchronizer.uploadInterval, true);
-                break;
-            default:
-                throw new IllegalArgumentException("Unknown SyncMode '" + syncMode + "'");
-        }
-    }
-
-    /**
-     * Schedules either a job handling the Upload (<code>LOCAL_TO_DROPBOX</code>)
-     * or Download (<code>DROPBOX_TO_LOCAL</code>) direction depending on
-     * <code>isUpload</code>.
-     * 
-     * @param interval the Trigger interval as cron expression
-     * @param isUpload
-     */
-    private void schedule(String interval, boolean isUpload) {
-        String direction = isUpload ? "Upload" : "Download";
-        try {
-            Scheduler sched = StdSchedulerFactory.getDefaultScheduler();
-            JobDetail job = newJob(SynchronizationJob.class).withIdentity(direction, DROPBOX_SCHEDULER_GROUP).build();
-
-            CronTrigger trigger = newTrigger().withIdentity(direction, DROPBOX_SCHEDULER_GROUP)
-                    .withSchedule(CronScheduleBuilder.cronSchedule(interval)).build();
-
-            logger.debug("Scheduled synchronization job (direction={}) with cron expression '{}'", direction, interval);
-            sched.scheduleJob(job, trigger);
-        } catch (SchedulerException e) {
-            logger.warn("Could not create synchronization job: {}", e.getMessage());
-        }
-    }
-
-    /**
-     * Delete all quartz scheduler jobs of the group <code>Dropbox</code>.
-     */
-    private void cancelAllJobs() {
-        try {
-            Scheduler sched = StdSchedulerFactory.getDefaultScheduler();
-            Set<JobKey> jobKeys = sched.getJobKeys(jobGroupEquals(DROPBOX_SCHEDULER_GROUP));
-            if (jobKeys.size() > 0) {
-                sched.deleteJobs(new ArrayList<JobKey>(jobKeys));
-                logger.debug("Found {} synchronization jobs to delete from DefaultScheduler (keys={})", jobKeys.size(),
-                        jobKeys);
-            }
-        } catch (SchedulerException e) {
-            logger.warn("Couldn't remove synchronization job: {}", e.getMessage());
-        }
-    }
-
-    /**
-     * A quartz scheduler job to execute the synchronization. There can be only
-     * one instance of a specific job type running at the same time.
-     */
-    @DisallowConcurrentExecution
-    public static class SynchronizationJob implements Job {
-
-        private final static JobKey UPLOAD_JOB_KEY = new JobKey("Upload", DROPBOX_SCHEDULER_GROUP);
-
-        @Override
-        public void execute(JobExecutionContext context) throws JobExecutionException {
-            boolean isUpload = UPLOAD_JOB_KEY.compareTo(context.getJobDetail().getKey()) == 0;
-
-            DropboxSynchronizer synchronizer = DropboxSynchronizer.instance;
-            if (synchronizer != null) {
-                try {
-                    DbxClient client = getClient(synchronizer);
-                    if (client != null) {
-                        if (isUpload) {
-                            synchronizer.syncLocalToDropbox(client);
-                        } else {
-                            synchronizer.syncDropboxToLocal(client);
-                        }
-                    } else {
-                        logger.info("Couldn't create Dropbox client. Most likely there has been no "
-                                + "access token found. Please restart the authentication process by"
-                                + " typing 'startAuthentication' on the OSGi console");
-                    }
-                } catch (Exception e) {
-                    logger.warn("Synchronizing data with Dropbox threw an exception", e);
-                }
-            } else {
-                logger.debug("DropboxSynchronizer instance hasn't been initialized properly!");
-            }
-        }
-
-        /**
-         * Creates and returns a new {@link DbxClient} initialized with the store access token.
-         * Returns {@code null} if no access token has been found.
-         * 
-         * @return a new {@link DbxClient} or <code>null</code> if no access token has been found.
-         */
-        private DbxClient getClient(DropboxSynchronizer synchronizer) {
-            String accessToken = synchronizer.readAccessToken();
-            if (StringUtils.isNotBlank(accessToken)) {
-                logger.debug("Creating new DbxClient");
-                return new DbxClient(requestConfig, accessToken);
-            }
-            return null;
-        }
-    }
-
-    static private String getUserDbxDataFolder() {
-        String progArg = System.getProperty("smarthome.userdata");
-        if (progArg != null) {
-            return progArg + File.separator + "dropbox";
-        } else {
-            return ".";
-        }
-    }
-
-    static private String getConfigDirFolder() {
-        String smartHomeProgArg = System.getProperty("smarthome.configdir");
-        String openHABProgArg = System.getProperty("openhab.configdir");
-        if (smartHomeProgArg != null) {
-            return smartHomeProgArg;
-        } else if (openHABProgArg != null) {
-            return openHABProgArg;
-        } else {
-            return ".";
-        }
-    }
-
-    private boolean isWindows() {
-        return (System.getProperty("os.name").toLowerCase().indexOf("win") >= 0);
     }
 }
